Break out type mapping logic
